requirements/base.txt
pip
setuptools
wheel
tqdm
pyyaml==5.3.1
numpy==1.19.4
pandas==1.1.3
bs4==0.0.1
anytree==2.8.0
matplotlib==3.3.3
spacy-nightly[transformers,lookups]==3.0.0rc2
https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.0.0a0/en_core_web_trf-3.0.0a0.tar.gz#egg=en_core_web_trf
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.0.0a0/en_core_web_lg-3.0.0a0.tar.gz#egg=en_core_web_lg
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.0.0a0/en_core_web_md-3.0.0a0.tar.gz#egg=en_core_web_md
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0a0/en_core_web_sm-3.0.0a0.tar.gz#egg=en_core_web_sm
